requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-bitfield==2.2.0
django-geo==0.8
django-js-asset==2.1.0
django-leaflet==0.29.0
django-mptt==0.15.0
django-separatedvaluesfield==0.6.0
django-timezone-field==6.0.1
django-uuidfield==0.5.0
djangorestframework==3.14.0
idna==3.4
lxml==4.9.3
psycopg2-binary==2.9.9
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
South==1.0.2
sqlparse==0.4.4
typing_extensions==4.8.0
unicodecsv==0.14.1
urllib3==2.0.7
